Refactor and add shortcut for rolling a : D, R and f5
