+++ lib/shortcuts/shortcuts.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+class DiceIntent extends Intent {
+  const DiceIntent();
+}
